move mergeBasket logic into its own hook
